+++ packages/core/src/config/index.ts
@@ -1,2 +1,2 @@
 export * from "./docs-config";
-export * from "./provider";
+export * from "./config.service";
